fix(slurm): Fix race condition in persistent connection termination
This fixes a race condition when using persistent connections
in `slurm_persist_conn_recv_server_fini` and `_service_connection`.
When a shutdown is signaled via reconfigure, there is a race between the call to
`pthread_detach` in `_service_connection` and `slurm_thread_join` in
`slurm_persist_conn_recv_server_fini`. This patch adds a guard around `pthread_detach`
to only run when we are not shutting down.

Co-Authored-by: Andrew Senetar <asenetar@coreweave.com>
Signed-off-by: Brad Beam <bbeam@coreweave.com>